+++ packages/effect/src/Effect.ts
@@ -3419,7 +3419,7 @@ export const updateService: {
 // -------------------------------------------------------------------------------------
 
 /**
- * The "do simulation" in allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
+ * The "do simulation" in Effect allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
  *
  * Here's how the do simulation works:
  *
@@ -3449,7 +3449,7 @@ export const updateService: {
 export const Do: Effect<{}> = effect.Do
 
 /**
- * The "do simulation" in allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
+ * The "do simulation" in Effect allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
  *
  * Here's how the do simulation works:
  *
@@ -3489,7 +3489,7 @@ export const bind: {
 } = effect.bind
 
 /**
- * The "do simulation" in allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
+ * The "do simulation" in Effect allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
  *
  * Here's how the do simulation works:
  *
@@ -3535,7 +3535,7 @@ const let_: {
 
 export {
   /**
-   * The "do simulation" in allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
+   * The "do simulation" in Effect allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
    *
    * Here's how the do simulation works:
    *
@@ -4029,6 +4029,12 @@ export const tap: {
   ) => [X] extends [Effect<infer _A1, infer E1, infer R1>] ? Effect<A, E | E1, R | R1>
     : [X] extends [PromiseLike<infer _A1>] ? Effect<A, E | Cause.UnknownException, R>
     : Effect<A, E, R>
+  <A, X, E1, R1>(
+    f: (a: NoInfer<A>) => Effect<X, E1, R1>,
+    options: { onlyEffect: true }
+  ): <E, R>(
+    self: Effect<A, E, R>
+  ) => Effect<A, E | E1, R | R1>
   <X>(
     f: NotFunction<X>
   ): <A, E, R>(
@@ -4036,18 +4042,34 @@ export const tap: {
   ) => [X] extends [Effect<infer _A1, infer E1, infer R1>] ? Effect<A, E | E1, R | R1>
     : [X] extends [PromiseLike<infer _A1>] ? Effect<A, E | Cause.UnknownException, R>
     : Effect<A, E, R>
+  <X, E1, R1>(
+    f: Effect<X, E1, R1>,
+    options: { onlyEffect: true }
+  ): <A, E, R>(
+    self: Effect<A, E, R>
+  ) => Effect<A, E | E1, R | R1>
   <A, E, R, X>(
     self: Effect<A, E, R>,
     f: (a: NoInfer<A>) => X
   ): [X] extends [Effect<infer _A1, infer E1, infer R1>] ? Effect<A, E | E1, R | R1>
     : [X] extends [PromiseLike<infer _A1>] ? Effect<A, E | Cause.UnknownException, R>
     : Effect<A, E, R>
+  <A, E, R, X, E1, R1>(
+    self: Effect<A, E, R>,
+    f: (a: NoInfer<A>) => Effect<X, E1, R1>,
+    options: { onlyEffect: true }
+  ): Effect<A, E | E1, R | R1>
   <A, E, R, X>(
     self: Effect<A, E, R>,
     f: NotFunction<X>
   ): [X] extends [Effect<infer _A1, infer E1, infer R1>] ? Effect<A, E | E1, R | R1>
     : [X] extends [PromiseLike<infer _A1>] ? Effect<A, E | Cause.UnknownException, R>
     : Effect<A, E, R>
+  <A, E, R, X, E1, R1>(
+    self: Effect<A, E, R>,
+    f: Effect<X, E1, R1>,
+    options: { onlyEffect: true }
+  ): Effect<A, E | E1, R | R1>
 } = core.tap
 
 /**
@@ -5958,6 +5980,8 @@ export declare namespace Tag {
    * @category models
    */
   export interface ProhibitedType {
+    Service?: `property "Service" is forbidden`
+    Identifier?: `property "Identifier" is forbidden`
     _op?: `property "_op" is forbidden`
     _tag?: `property "_tag" is forbidden`
     of?: `property "of" is forbidden`