+++ packages/platform/src/HttpApp.ts
@@ -21,9 +21,11 @@ import * as internalMiddleware from "./internal/httpMiddleware.js"
  * @since 1.0.0
  * @category models
  */
-export interface HttpApp<A = ServerResponse.HttpServerResponse, E = never, R = never>
-  extends Effect.Effect<A, E, R | ServerRequest.HttpServerRequest>
-{}
+export type HttpApp<A = ServerResponse.HttpServerResponse, E = never, R = never> = Effect.Effect<
+  A,
+  E,
+  R | ServerRequest.HttpServerRequest
+>
 
 /**
  * @since 1.0.0