+++ packages/schema/src/AST.ts
@@ -227,6 +227,18 @@ export const ParseIssueTitleAnnotationId = Symbol.for("@effect/schema/annotation
  */
 export const ParseOptionsAnnotationId = Symbol.for("@effect/schema/annotation/ParseOptions")
 
+/**
+ * @category annotations
+ * @since 0.70.1
+ */
+export type DecodingFallbackAnnotation<A> = (issue: ParseIssue) => Effect<A, ParseIssue>
+
+/**
+ * @category annotations
+ * @since 0.70.1
+ */
+export const DecodingFallbackAnnotationId = Symbol.for("@effect/schema/annotation/DecodingFallback")
+
 /** @internal */
 export const SurrogateAnnotationId = Symbol.for("@effect/schema/annotation/Surrogate")
 
@@ -364,6 +376,14 @@ export const getParseIssueTitleAnnotation = getAnnotation<ParseIssueTitleAnnotat
  * @since 0.68.3
  */
 export const getParseOptionsAnnotation = getAnnotation<ParseOptions>(ParseOptionsAnnotationId)
+
+/**
+ * @category annotations
+ * @since 0.70.1
+ */
+export const getDecodingFallbackAnnotation = getAnnotation<DecodingFallbackAnnotation<unknown>>(
+  DecodingFallbackAnnotationId
+)
 
 /** @internal */
 export const getSurrogateAnnotation = getAnnotation<SurrogateAnnotation>(SurrogateAnnotationId)