+++ packages/sql-mysql2/test/utils.ts
@@ -1,0 +1,33 @@
+import { MysqlClient } from "@effect/sql-mysql2"
+import type { StartedMySqlContainer } from "@testcontainers/mysql"
+import { MySqlContainer } from "@testcontainers/mysql"
+import { Config, Context, Data, Effect, Layer, Redacted } from "effect"
+
+export class ContainerError extends Data.TaggedError("ContainerError")<{
+  cause: unknown
+}> {}
+
+export class MysqlContainer extends Context.Tag("test/MysqlContainer")<
+  MysqlContainer,
+  StartedMySqlContainer
+>() {
+  static Live = Layer.scoped(
+    this,
+    Effect.acquireRelease(
+      Effect.tryPromise({
+        try: () => new MySqlContainer("mysql:lts").start(),
+        catch: (cause) => new ContainerError({ cause })
+      }),
+      (container) => Effect.promise(() => container.stop())
+    )
+  )
+
+  static ClientLive = Layer.unwrapEffect(
+    Effect.gen(function*(_) {
+      const container = yield* _(MysqlContainer)
+      return MysqlClient.layer({
+        url: Config.succeed(Redacted.make(container.getConnectionUri()))
+      })
+    })
+  ).pipe(Layer.provide(this.Live))
+}